fix(frontend): make camper detail panel a frozen slide-in overlay
Opening the camper detail panel used to reflow the bunk board: it
measured the panel's overlap onto the board, trimmed the right margin,
and dropped a grid column on mid-width screens. Every bunk card re-wrapped
into new rows and an auto-pan re-centred the clicked bunk, so staff lost
the bunk they were working on.

Make the panel a plain fixed right-side slide-in overlay that never moves
the board — no margin trim, no column drop, no auto-pan. The board stays
frozen; the panel simply floats over the right edge while open. This
matches the friend-groups (lock group) panel, which was already a plain
left-side overlay.

Removes computeBoardReflow / getBunkGridClass / bunkAutoPan and their
tests; keeps getBoardBottomPaddingClass (unrelated hub clearance).
