+++ frontend/src/components/BunkingBoardByArea.tsx
@@ -1,4 +1,4 @@
-import { useState, useTransition, useEffect, useCallback, useRef, lazy, Suspense } from 'react'
+import { useState, useTransition, useEffect, useCallback, lazy, Suspense } from 'react'
 import type { DragEndEvent, DragStartEvent } from '@dnd-kit/core'
 import {
   DndContext,
@@ -36,13 +36,7 @@ import { Home } from 'lucide-react'
 import { isAgSession } from '../utils/sessionTypePredicates'
 import { getEffectivelyUnassignedCampers } from './bunkingBoardHelpers'
 import { shouldKeepPanelsOpen } from '../utils/clickoutsidePredicate'
-import {
-  computeBoardReflow,
-  getBoardBottomPaddingClass,
-  getBunkGridClass,
-  type BoardReflowResult,
-} from '../utils/bunkBoardLayout'
-import { autoPanToBunk } from '../utils/bunkAutoPan'
+import { getBoardBottomPaddingClass } from '../utils/bunkBoardLayout'
 
 interface BunkingBoardByAreaProps {
   sessionId: string
@@ -103,20 +97,6 @@ export default function BunkingBoardByArea(props: BunkingBoardByAreaProps) {
   const currentYear = useYear()
   const { hasPermission } = usePermissions()
 
-  // Ref for the board scroll container — used by autoPanToBunk to scroll the
-  // selected camper's bunk card into view when the detail panel opens.
-  const boardScrollRef = useRef<HTMLDivElement>(null)
-  // Measured reflow state for the camper-detail panel (see computeBoardReflow).
-  // marginRightPx trims the board by the panel's actual overlap; dropColumn /
-  // didReflow drive the grid column count and whether auto-pan should fire.
-  const [reflow, setReflow] = useState<BoardReflowResult>({
-    marginRightPx: 0,
-    dropColumn: false,
-    didReflow: false,
-  })
-  // Right margin currently applied to the board wrapper, so a re-measure can
-  // recover the board's *natural* right edge (rect.right + appliedMargin).
-  const appliedMarginRef = useRef(0)
   const canManage = hasPermission(Permission.BUNKING_MANAGE)
 
   // Get lock group context for action bar and pending camper management
@@ -605,47 +585,6 @@ export default function BunkingBoardByArea(props: BunkingBoardByAreaProps) {
     }
   }, [isAnyPanelOpen, handleGlobalClick])
 
-  // Measure how the camper-detail panel should reflow the board. The board sits
-  // in a centered max-w-7xl container, so on wide screens the fixed panel floats
-  // over the background gutter and nothing needs to move; on narrower screens the
-  // board is trimmed from the right by the panel's actual overlap only. We recover
-  // the board's natural right edge by adding back the margin we already applied.
-  const measureReflow = useCallback(() => {
-    const el = boardScrollRef.current
-    if (!el || !selectedCamperId) {
-      appliedMarginRef.current = 0
-      setReflow({ marginRightPx: 0, dropColumn: false, didReflow: false })
-      return
-    }
-    const rect = el.getBoundingClientRect()
-    const result = computeBoardReflow({
-      boardNaturalLeft: rect.left,
-      boardNaturalRight: rect.right + appliedMarginRef.current,
-      viewportWidth: window.innerWidth,
-    })
-    appliedMarginRef.current = result.marginRightPx
-    setReflow(result)
-  }, [selectedCamperId])
-
-  // Re-measure on selection change, and while the panel is open on window resize.
-  // Resize re-measures the margin/columns but never pans (handled separately).
-  useEffect(() => {
-    measureReflow()
-    if (!selectedCamperId) return
-    window.addEventListener('resize', measureReflow)
-    return () => window.removeEventListener('resize', measureReflow)
-  }, [selectedCamperId, measureReflow])
-
-  // Auto-pan: scroll the selected camper's bunk into view — but ONLY when the
-  // board actually reflowed (a column dropped). On wide screens the board does
-  // not move, so an auto-pan would be an unwanted jump. autoPanToBunk also
-  // no-ops when the bunk is already visible or the camper is unassigned.
-  useEffect(() => {
-    if (!selectedCamperId || !reflow.didReflow) return
-    const selected = campers.find((c) => String(c.person_cm_id) === selectedCamperId)
-    autoPanToBunk(selected?.assigned_bunk_cm_id ?? null, boardScrollRef.current)
-  }, [selectedCamperId, reflow.didReflow, campers])
-
   return (
     <>
       <DndContext
@@ -655,22 +594,16 @@ export default function BunkingBoardByArea(props: BunkingBoardByAreaProps) {
         onDragEnd={handleDragEnd}
         onDragCancel={handleDragCancel}
       >
-        {/* Main bunks area — trimmed from the right by the panel's actual overlap
-            when the camper-detail panel is open (see computeBoardReflow); gets
-            bottom padding to clear the friend-groups hub / action bar (#1630). */}
+        {/* Main bunks area. The camper-detail panel is a plain right-side slide-in
+            overlay (fixed), so the board never moves when it opens — staff keep the
+            bunk they're working on in place. Only gets bottom padding to clear the
+            friend-groups hub / action bar (#1630). */}
         <div
           data-board-wrapper
-          ref={boardScrollRef}
-          className={[
-            'transition-[margin] duration-200 ease-out',
-            getBoardBottomPaddingClass(
-              isLockGroupUiActive,
-              isLockGroupUiActive && pendingCampers.length > 0
-            ),
-          ]
-            .filter(Boolean)
-            .join(' ')}
-          style={reflow.marginRightPx ? { marginRight: `${reflow.marginRightPx}px` } : undefined}
+          className={getBoardBottomPaddingClass(
+            isLockGroupUiActive,
+            isLockGroupUiActive && pendingCampers.length > 0
+          )}
         >
           {/* Lock all / Unlock all — visible only to managers in non-production mode */}
           {canManage && !isProductionMode && onLockAll && onUnlockAll && (
@@ -700,7 +633,7 @@ export default function BunkingBoardByArea(props: BunkingBoardByAreaProps) {
           ) : (
             <div
               data-bunk-grid
-              className={getBunkGridClass(reflow.dropColumn)}
+              className="grid grid-cols-1 gap-3 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4"
               style={{ contain: 'layout style' }}
               onClick={handleBoardClick}
             >
